Attachment plugin tests
