+++ tests/conftest.py
@@ -1,4 +1,11 @@
+import json
+from typing import Tuple
+
 import pytest
+import requests
+from kinto_http import Client
+from requests.adapters import HTTPAdapter
+from urllib3.util.retry import Retry
 
 
 DEFAULT_SERVER = "http://localhost:8888/v1"
@@ -85,3 +92,80 @@ def collection(request):
 @pytest.fixture(scope="session")
 def reset(request):
     return request.config.getoption("--reset")
+
+
+@pytest.fixture
+def get_clients(
+    flush_server,
+    server: str,
+    auth: Tuple[str, str],
+    editor_auth: Tuple[str, str],
+    reviewer_auth: Tuple[str, str],
+    bucket: str,
+    collection: str,
+) -> Tuple[Client, Client, Client]:
+    """Pytest fixture for creating Kinto Clients used for tests.
+
+    Args:
+        server (str): Kinto server (in form 'http(s)://<host>:<port>/v1')
+        auth (Tuple[str, str]): Basic authentication where auth[0]=user, auth[1]=pass
+        editor_auth (str): Basic authentication for editor
+        reviewer_auth (str): Basic authentication for reviewer
+        bucket (str): Source bucket
+        collection (str): Source collection
+
+    Returns:
+        Tuple[Client, Client, Client]: User client, editor client, reviewer client
+    """
+
+    request_session = requests.Session()
+    retries = Retry(total=5, backoff_factor=0.1, status_forcelist=[500, 502, 503, 504])
+    request_session.mount(
+        f"{server.split('://')[0]}://", HTTPAdapter(max_retries=retries)
+    )
+
+    # check if user already exists before creating
+    r = request_session.get(f"{server}/", auth=auth)
+    if "user" not in r.json():
+        create_account_url = f"{server}/accounts/{auth[0]}"
+        assert request_session.put(
+            create_account_url,
+            data=json.dumps({"data": {"password": auth[1]}}),
+            headers={"Content-Type": "application/json"},
+        )
+
+    client = Client(
+        server_url=server,
+        auth=auth,
+        bucket=bucket,
+        collection=collection,
+        retry=5,
+    )
+
+    if editor_auth is None:
+        editor_auth = auth
+
+    if reviewer_auth is None:
+        reviewer_auth = auth
+
+    editor_client = Client(
+        server_url=server,
+        auth=editor_auth,
+        bucket=bucket,
+        collection=collection,
+        retry=5,
+    )
+    reviewer_client = Client(
+        server_url=server,
+        auth=reviewer_auth,
+        bucket=bucket,
+        collection=collection,
+        retry=5,
+    )
+
+    return client, editor_client, reviewer_client
+
+
+@pytest.fixture
+def flush_server(server: str):
+    assert requests.post(f"{server}/__flush__")
