+++ tests/integration_test.py
@@ -1,18 +1,14 @@
-import json
 import os
 from typing import Tuple
 
-import pytest
 import requests
 from kinto_http import Client
-from requests.adapters import HTTPAdapter
-from urllib3.util.retry import Retry
 
 
-def test_history_plugin(get_clients: Tuple[Client, Client, Client, str]):
-    client, _, _, _ = get_clients
+def test_history_plugin(get_clients: Tuple[Client, Client, Client]):
+    client, _, _ = get_clients
     client.create_bucket(id="blog")
-    client.create_collection(id="articles", bucket="blog", if_not_exists=True)
+    client.create_collection(id="articles", bucket="blog")
     history = client.get_history(bucket="blog")
 
     assert history
@@ -23,7 +19,7 @@ def test_history_plugin(get_clients: Tuple[Client, Client, Client, str]):
     assert "blog" in history[1]["bucket_id"]
 
 
-def test_email_plugin(get_clients: Tuple[Client, Client, Client, str]):
+def test_email_plugin(get_clients: Tuple[Client, Client, Client]):
     # remove any existing .eml files in mail directory
     try:
         for file in os.listdir("mail"):
@@ -31,7 +27,7 @@ def test_email_plugin(get_clients: Tuple[Client, Client, Client, str]):
     except FileNotFoundError:
         pass
 
-    client, editor_client, _, _ = get_clients
+    client, editor_client, _ = get_clients
     client.create_bucket(id="source")
     client.create_collection(id="email", bucket="source")
     client.patch_bucket(
@@ -67,78 +63,34 @@ def test_email_plugin(get_clients: Tuple[Client, Client, Client, str]):
         assert mail_contents.find("To: me@you.com") >= 0
 
 
-@pytest.fixture
-def get_clients(
-    flush_server,
-    server: str,
-    auth: Tuple[str, str],
-    editor_auth: Tuple[str, str],
-    reviewer_auth: Tuple[str, str],
-    bucket: str,
-    collection: str,
-) -> Tuple[Client, Client, Client, str]:
-    """Pytest fixture for creating Kinto Clients used for tests.
+def test_attachment_plugin_new_record(
+    get_clients: Tuple[Client, Client, Client], server: str
+):
+    client, _, _ = get_clients
+    client.create_bucket(id="blog")
+    client.create_collection(id="articles", bucket="blog")
 
-    Args:
-        server (str): Kinto server (in form 'http(s)://<host>:<port>/v1')
-        auth (Tuple[str, str]): Basic authentication where auth[0]=user, auth[1]=pass
-        editor_auth (str): Basic authentication for editor
-        reviewer_auth (str): Basic authentication for reviewer
-        bucket (str): Source bucket
-        collection (str): Source collection
-
-    Returns:
-        Tuple[Client, Client, Client, str]: User client, editor client, reviewer client, server
-    """
-
-    request_session = requests.Session()
-    retries = Retry(total=5, backoff_factor=0.1, status_forcelist=[500, 502, 503, 504])
-    request_session.mount(
-        f"{server.split('://')[0]}://", HTTPAdapter(max_retries=retries)
-    )
-
-    # check if user already exists before creating
-    r = request_session.get(f"{server}/", auth=auth)
-    if "user" not in r.json():
-        create_account_url = f"{server}/accounts/{auth[0]}"
-        assert request_session.put(
-            create_account_url,
-            data=json.dumps({"data": {"password": auth[1]}}),
-            headers={"Content-Type": "application/json"},
+    with open("kinto-logo.svg", "rb") as attachment:
+        assert requests.post(
+            f"{server}/buckets/blog/collections/articles/records/logo/attachment",
+            files={"attachment": attachment},
+            auth=client.session.auth,
         )
 
-    client = Client(
-        server_url=server,
-        auth=auth,
-        bucket=bucket,
-        collection=collection,
-        retry=5,
+
+def test_attachment_plugin_existing_record(
+    get_clients: Tuple[Client, Client, Client], server: str
+):
+    client, _, _ = get_clients
+    client.create_bucket(id="blog")
+    client.create_collection(id="articles", bucket="blog")
+    client.create_record(
+        id="logo", bucket="blog", collection="articles", data={"type": "logo"}
     )
 
-    if editor_auth is None:
-        editor_auth = auth
-
-    if reviewer_auth is None:
-        reviewer_auth = auth
-
-    editor_client = Client(
-        server_url=server,
-        auth=editor_auth,
-        bucket=bucket,
-        collection=collection,
-        retry=5,
-    )
-    reviewer_client = Client(
-        server_url=server,
-        auth=reviewer_auth,
-        bucket=bucket,
-        collection=collection,
-        retry=5,
-    )
-
-    return client, editor_client, reviewer_client, server
-
-
-@pytest.fixture
-def flush_server(server: str):
-    assert requests.post(f"{server}/__flush__")
+    with open("kinto-logo.svg", "rb") as attachment:
+        assert requests.post(
+            f"{server}/buckets/blog/collections/articles/records/logo/attachment",
+            files={"attachment": attachment},
+            auth=client.session.auth,
+        )
